Chaged naming to snake case and made names clearer
