--- processes/fulfillURLManifest.py
@@ -1,168 +1,0 @@
-import json
-import os
-import copy
-import logging
-from botocore.exceptions import ClientError
-
-from .core import CoreProcess
-from datetime import datetime, timedelta, timezone
-from model import Link
-from logger import createLog
-
-logger = createLog(__name__)
-
-class FulfillProcess(CoreProcess):
-
-    def __init__(self, *args):
-        super(FulfillProcess, self).__init__(*args[:4])
-
-        self.fullImport = self.process == 'complete' 
-        self.startTimestamp = None
-
-        # Connect to database
-        self.generateEngine()
-        self.createSession()
-
-        # S3 Configuration
-        self.s3Bucket = os.environ['FILE_BUCKET']
-        self.host = os.environ['DRB_API_HOST']
-        self.prefix = 'manifests/UofM/'
-        self.createS3Client()
-
-    def runProcess(self):
-        if self.process == 'daily':
-            startTimeStamp = datetime.now(timezone.utc) - timedelta(days=1)
-            self.getManifests(startTimeStamp)
-        elif self.process == 'complete':
-            self.getManifests()
-        elif self.process == 'custom':
-            timeStamp = self.ingestPeriod
-            startTimeStamp = datetime.strptime(timeStamp, '%Y-%m-%dT%H:%M:%S')
-            self.getManifests(startTimeStamp)
-
-    def getManifests(self, startTimeStamp=None):
-
-        '''Load batch of LA works based on startTimeStamp'''
-
-        batches = self.load_batches(self.prefix, self.s3Bucket)
-        if startTimeStamp:
-            #Using JMESPath to extract keys from the JSON batches
-            filtered_batch_keys = batches.search(f"Contents[?to_string(LastModified) > '\"{startTimeStamp}\"'].Key")
-            for key in filtered_batch_keys:
-                metadataObject = self.s3Client.get_object(Bucket=self.s3Bucket, Key= f'{key}')
-                self.update_manifest(metadataObject, self.s3Bucket, key)
-        else:
-            for batch in batches:
-                for content in batch['Contents']:
-                    key = content['Key']
-                    metadataObject = self.s3Client.get_object(Bucket=self.s3Bucket, Key= f'{key}')
-                    self.update_manifest(metadataObject, self.s3Bucket, key)
-
-    def update_manifest(self, metadataObject, bucketName, currKey):
-
-        metadataJSON = json.loads(metadataObject['Body'].read().decode("utf-8"))
-        metadataJSONCopy = copy.deepcopy(metadataJSON)
-
-        #Checking if record asscociated with manifest is public domain or in copyright
-        checkRightsStatus = 'in copyright'
-        checkRightsStatus = self.checkMetaData(metadataJSON, checkRightsStatus)
-
-        if checkRightsStatus == 'in copyright':
-        
-            counter = 0
-        
-            metadataJSON, counter = self.linkFulfill(metadataJSON, counter)
-            metadataJSON, counter = self.readingOrderFulfill(metadataJSON, counter)
-            metadataJSON, counter = self.resourceFulfill(metadataJSON, counter)
-            metadataJSON, counter = self.tocFulfill(metadataJSON, counter)
-
-            if counter >= 4: 
-                for link in metadataJSON['links']:
-                    self.fulfillFlagUpdate(link)
-
-            self.closeConnection()
-
-            if metadataJSON != metadataJSONCopy:
-                try:
-                    fulfillManifest = json.dumps(metadataJSON, ensure_ascii = False)
-                    return self.s3Client.put_object(
-                        Bucket=bucketName, 
-                        Key=currKey, 
-                        Body=fulfillManifest, ACL= 'public-read', 
-                        ContentType = 'application/json'
-                    )
-                except ClientError as e:
-                    logging.error(e)
-
-    def linkFulfill(self, metadataJSON, counter):
-        for link in metadataJSON['links']:
-            fulfillLink, counter = self.fulfillReplace(link, counter)
-            link['href'] = fulfillLink
-
-        return (metadataJSON, counter)
-            
-    def readingOrderFulfill(self, metadataJSON, counter):
-        for readOrder in metadataJSON['readingOrder']:
-            fulfillLink, counter = self.fulfillReplace(readOrder, counter)
-            readOrder['href'] = fulfillLink
-
-        return (metadataJSON, counter)
-
-    def resourceFulfill(self, metadataJSON, counter):
-        for resource in metadataJSON['resources']:
-            fulfillLink, counter = self.fulfillReplace(resource, counter)
-            resource['href'] = fulfillLink
-
-        return (metadataJSON, counter)
-
-    def tocFulfill(self, metadataJSON, counter): 
-
-        '''
-        The toc dictionary has no "type" key like the previous dictionaries 
-        therefore the 'href' key is evaluated instead
-        '''
-
-        for toc in metadataJSON['toc']:
-            if 'pdf' in toc['href'] \
-                or 'epub' in toc['href']:
-                    for link in self.session.query(Link) \
-                        .filter(Link.url == toc['href'].replace('https://', '')):
-                            counter += 1
-                            toc['href'] = f'https://{self.host}/fulfill/{link.id}'
-
-        return (metadataJSON, counter)
-
-    def fulfillReplace(self, metadata, counter):
-        if metadata['type'] == 'application/pdf' or metadata['type'] == 'application/epub+zip' \
-            or metadata['type'] == 'application/epub+xml':
-                for link in self.session.query(Link) \
-                    .filter(Link.url == metadata['href'].replace('https://', '')):
-                            counter += 1            
-                            metadata['href'] = f'https://{self.host}/fulfill/{link.id}'
-
-        return (metadata['href'], counter)
-    
-    def fulfillFlagUpdate(self, metadata):
-        if metadata['type'] == 'application/webpub+json':
-            for link in self.session.query(Link) \
-                .filter(Link.url == metadata['href'].replace('https://', '')):   
-                        print(link)
-                        print(link.flags)
-                        if 'fulfill_limited_access' in link.flags.keys():
-                            if link.flags['fulfill_limited_access'] == False:
-                                newLinkFlag = dict(link.flags)
-                                newLinkFlag['fulfill_limited_access'] = True
-                                link.flags = newLinkFlag
-                                self.commitChanges()
-                        
-    def checkMetaData(self, checkRightsStatus, metadataJSON):
-         for link in metadataJSON['links']:
-            if metadataJSON['type'] == 'application/webpub+json':
-                for link in self.session.query(Link) \
-                    .filter(Link.url == metadataJSON['href'].replace('https://', '')):   
-                        if 'fulfill_limited_access' not in link.flags.keys():
-                            checkRightsStatus = 'public domain'
-                        return checkRightsStatus
-
-class FulfillError(Exception):
-    pass
